# Add CTEs and optimize fct_daily_rt_feed_validation_notices